Move tests to PartitionTest
